Feat: 실시간 종목 랭킹 시스템 (등락률/거래대금/거래량)
Redis Sorted Set 기반 실시간 랭킹 구현
- 체결 시마다 자동 업데이트 (incrementScore/add)
- InfluxDB 스케줄러 제거 → 성능 개선
- Read/Write 서비스 분리 설계

Added:
- MarketRankWriterService (랭킹 쓰기)
- MarketRankReaderService (랭킹 조회)
- TradingRankController (API 엔드포인트)

Removed:
- TradingRankScheduler (불필요)
- TradingRankService (InfluxDB 집계 로직)

Refs: #128

+++ marketdata/src/main/java/com/beyond/MKX/domain/orderbook/dto/OrderStatusEventDTO.java
@@ -29,6 +29,6 @@ public class OrderStatusEventDTO {
     private long avgFillPrice;        // 체결 평균가
     private long lastFillPrice;       // 마지막 체결가
     private long limitPrice;          // 지정가
-    private BigDecimal filledQuantity;// 누적 체결 수량
+    private long filledQuantity;// 누적 체결 수량
     private long notional;            // 총 거래대금
 }
